+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.9]
+
+### Added
+
+- `hsha/threaded_selector` 增加 `backlog` 配置, 避免连接过多时, 服务端拒绝/重置连接
+- 增加 `ThriftThreadFactory`; 所有thrift工作线程按 `thrift-[server]-thread-%d` 命名
+
+### Changed
+
+- apache thrift 版本升级: 0.9.3 -> 0.19.0
+
+## [1.0.9]
+
+### Changed
+
+- ThriftServer 支持 Spring AOP 切面
+
 ## [1.0.8]
 
 ### Changed